requirements.txt
requests
beautifulsoup4
colorama
tabulate
python-whois
dnspython
